fix: fix utils
