+++ src/index.ts
@@ -13,11 +13,13 @@ async function start() {
 
     const GLOBAL_DATA: IUser | [] = []
 
-    const server = createServer()
-    server.on('request', (req, res) => {
-        console.log('Request path' + req.url);
-        router(req, res, GLOBAL_DATA)         
+    const server = createServer((req, res) => {
+        router(req, res, GLOBAL_DATA) 
     })
+    // server.on('request', (req, res) => {
+    //     console.log('Request path' + req.url);
+    //             
+    // })
 
     server.listen(PORT, 'localhost', () => {
         console.log(`App has been started on port ${PORT}...`)});
