+++ src/utils/get-name-from-path.ts
@@ -1,0 +1,6 @@
+import {parse} from 'path'
+
+export const getNameFromPath = async (url: string) => {
+    const {name} = parse(url) 
+    return name
+}
